Build 000, PSDto3D initial version, CDRIN source snapshot
Snapshot of the original PSDto3D source delivered by CDRIN to E.D. Films in November 2018, except for the following:

- Original source was arranged into six separate Git repositories.  These are now combined as subfolders of a single repository.

- Technical documentation was not originally under source control, now included in the 'Technical_Documentation' folder.

- Output folder from psd2m_installer is removed, which contained a generated installer binary.

- Project files for Microsoft Visual Studio (.vcxproj and .sln) were not originally under source control, now included in the 'projects' folder.  Originally, these were generated at compile time using CMake.  Since Visual Studio is the primary development software for this project, and only Windows is supported, with no support planned for Mac or Linux, CMake is no longer used in the build pipeline.  The project files are now under source control, and no longer generated at compile time.

- Original project files for Microsoft Visual Studio, generated by CDRIN for Visual Studio 2017, have been lost.  Newly generated project files for Visual Studio 2019 are now included.  These use Windows SDK Version 10.0.19041.0, and Platform Toolkit Visual Studio 2019 v142.  As such, the plugin built with these project files is not binary identical to the plugin delivered by CDRIN.  But, the newly recompiled plugin has been tested in Maya 2017, 2018 and 2019.  Running the plugin requires installing the Visual Studio 2019 Redistributable package.

### Copyright and licensing

The source code is owned by, and proprietary to, E.D. Films and is not open source.  The original code from CDRIN contained LICENSE.md files indicating a GNU General Public License.  But this was simply due to a Git template used, and is not applicable because ownership was fully transferred to E.D. Films.  The original LICENSE.md files are still included here; this is for the sake of preserving an exact snapshot of the original source code.

### Supported Maya versions

The plugin originally supported Maya 2016, 2017, and 2018, and CDRIN delivered an installer for these versions.  This was changed to Maya 2017, 2018 and 2019 in the first public release.  Support was dropped for Maya 2016, which used an older Qt 4.0 instead of Qt 5.0, causing difficulties in compilation.  Also, as Maya 2016 is no longer available for download, it was not possible to produce a newly generated project file, or test the plugin for this version.

### File layout

- projects
Newly generated project files for Visual Studio 2019.

- psd2m_extension
Photoshop extension.  The shipped binary is in .zxp format, and the source code is in a javascript .jsx file.  The source is signed using the ZXPSignCmd tool and the final .zxp is simply by zipping the signed file into a compressed archive and renaming to .zxp.  The installer for the extension is Adobe's proprietary ExMan installer tool, which is included with, and invoked by, the main PSDto3D installer.

- psd2m_installer
Installer script.  The source code script is in .iss format.  This generates the installer when used with InnoSetup 5.6.1, which is freeware.  The InnoSetup software is not currently included here;  this is for the sake of preserving an exact snapshot of the original source code, which did not include it.

- psd2m_maya_plugin
The main project in C++, which compiles to the Maya plugin.

- psd2m_mesh_generation
Helper project in C++, which compiles to a .lib static library.  Provides algorithms for the Mesh Generation modes, including Vector mode, and support for Influence layers.

- psd2m_psd_reader
Helper project in C++, which compiles to a .lib static library.  Provides support for parsing the .psd file format.

- psd2m_util
Helper project in C++, which compiles to a .lib static library.  Provides basic math helpers and string routines.

- Technical_Documentation
Original information and documentation produced by CDRIN.

- README.md files are not updated or deleted; this is for the sake of preserving an exact snapshot of the original source code, which was arranged into six separate Git repositories.  However, a new CHANGELOG.md (this document) is added at the repository root.

- CONTRIBUTORS.md files are not updated or deleted; this is for the sake of preserving an exact snapshot of the original source code, which was arranged into six separate Git repositories.  However, a consolidated CONTRIBUTORS.md file is added at the repository root.

- LICENSE.md files are outdated as explained above

### Known Issues

- Installing the Visual Studio 2019 Redistributable package is a prerequisite to running the plugin, assuming the plugin is built with the Visual Studio 2019 compiler and using newly generated project files.  The package is available at https://aka.ms/vs/17/release/vc_redist.x64.exe

- The Maya shelf is not included here; this is for the sake of preserving an exact snapshot of the original source code, which did not include it.

- The following problems were noted in testing.  These might be original problems which existed in the build delivered by CDRIN.  Not diagnosed.
-- Attempting to use the Maya shelf from a recent build, with the plugin path edited, did not work in Maya2017 or Maya2018, producing a blank shelf in the UI, but this did work as expected in Maya 2019.
-- Linear mode is not working, producing an empty object in the scene.  But Vector mode generation and PNG generation both work as expected.
-- Texture is sometimes not connected to the material when a mesh is generated.
-- Generating meshes for multiple layers at once sometimes crashes.
-- Installer builder not tested.
-- Photoshop extension builder not tested.