PSDtoUnreal build script
Support to build and package the PSDtoUnreal plugin via the new script build_deployment_unreal.bat.  Note this script must be edited for each future version of Unreal as it embeds the BuildId number which depends on Unreal version.

PSDtoUnreal bugfix for issue where the export buttons were not working (Mesh, PNGs, Mesh+PNGs); internally the FileExportName could sometimes be an empty string if stored as empty in the parameters.json file, which in turn caused the export to fail.  The bug was caused by previous refactoring in the PSDtoFBX codebase.  This is fixed so FileExportName is always set to a non-empty value in PSDtoMaya and PSDtoUnreal, so the export should never fail for this reason.