PSDtoUnreal fix atlas error on reload after layer reordering
Fix for a workflow issue found in testing.  If some layers were assigned to a texture atlas, and then the PSD file was modified to add a new layer above those, and then Reload was used in PSDtoUnreal to load the updated PSD, the atlas would be garbled upon output and contain the wrong layers.  Internally, the Reload command would load the atlas settings from the JSON parameters file, referring to layers by index number, but these indices were incorrect as the new layer caused indices below it to increase.  The system would not adjust the atlas layer indices accordingly.  This is fixed so that, during reload, the parameters for each layer in the JSON are matched by name, not by index, to layers in the PSD, and indices for texture atlases are patched accordingly.

This also adds support for duplicate layer names, so that, for example, if a scene has two layers both named "shrub" with different settings, they are both saved to JSON and loaded back to the first two layers name "shrub", in the same order upon reload.  However, if the order of such layers is swapped in Photoshop, this will not be detected, and the system will apply settings to layers in the original (out of date) order.  Nevertheless, this update allows more robust support for adding and reordering layers during workflow iteration, especially when all layers have unique names.