Updated Photoshop extension signing key
Photoshop extension signing key is now updated, to resolve errors reported through customer support.  Previously, the extension would fail to load and would produce an error dialog in Photoshop.  This is fixed so the extension should run safely for four years from the build's release date.  No other fixes or changes.

Internally, the extension is now self-signed, with a new key generated every build by the build script.  Previously, all builds re-used a single key which expired in May 2023.  Using a self-signed key should allow the extension to run in Photoshop without signing errors for four years from the date of build.

This is a candidate final build of PSDtoMaya and PSDtoFBX to be available for the general public, before development shifts to a consulting business-to-business model.  The version number is increased to 1.6.5, from previous 1.6.4.