PSDtoFBX support for VS2022, and fix window position issue
Fix for build issue with PSDtoFBX in VS2022 (Visual Studio 2022).  Updated project file to specify VS2019 platform; previously unspecified.  Note this means that building PSDtoFBX requires the VS2019 platform toolset is installed, or selected as an option during VS2022 install.

Fix for issues where PSDto3D window sometimes appeared offscreen when first displayed, and was too tall vertically to fix on screen.  The window is now initially centered on the primary monitor, and its size is slightly reduced in the vertical dimension, so it should fit entirely onscreen by default.  Tested with a two-monitor setup at 1920x1080 and at 3840x2160.